// constants/enums.ts
export enum Status {
	ALL = "all",
	ALIVE = "alive",
	DEAD = "dead",
	UNKNOWN = "unknown",
}
export enum Gender {
	ALL = "all",
	FEMALE = "female",
	MALE = "male",
	UNKNOWN = "unknown",
	GENDERLESS = "genderless",
}
export enum Pages {
	CHARACTER = "character",
	LOCATIONS = "locations",
	EPISODES = "episodes",
}